Move BaseIntegrator to operation.py.
Also fix some pre-commit errors.
