+++ hoomd/CMakeLists.txt
@@ -347,7 +347,6 @@ set(files box.py
           device.py
           __init__.py
           error.py
-          integrate.py
           operation.py
           operations.py
           pytest_plugin_validate.py
